leave drag-stuff-define-keys up to the user
Putting a call to drag-stuff-define-keys in the body of code that runs
whenever drag-stuff-mode is enabled/disabled makes it overly complicated
to customize drag-stuff-mode-map when using drag-stuff-global-mode. The
user's keybindings refuse to take hold unless they're wrapped in a
function that's added to drag-stuff-mode-hook.

For example:
  (defun my/bind-drag-stuff-keys ()
    ;; unset the defaults
    (define-key drag-stuff-mode-map (drag-stuff--kbd 'up) nil)
    ...
    ;; apply my bindings
    (define-key drag-stuff-mode-map (kbd "C-M-k") 'drag-stuff-up)
    ...)

  (add-hook 'drag-stuff-mode-hook
	    'my/bind-drag-stuff-keys)

  (drag-stuff-global-mode t)

Without resorting to the above use of hooks, any custom bindings get
clobbered by drag-stuff-define-keys.

Resolve this issue by removing the call to drag-stuff-define-keys from
the body of the minor mode definition and adding instructions to the
readme as to how users may activate the suggested keybindings.